+++ leads_gpio/__init__.py
@@ -7,6 +7,6 @@ if not _find_spec("pynmea2"):
 if not _find_spec("serial"):
     raise ImportError("Please install `pyserial` to run this module\n>>>pip install pyserial")
 
-from leads_raspberry_pi.gps_receiver import *
-from leads_raspberry_pi.led import *
-from leads_raspberry_pi.led_group import *
+from leads_gpio.gps_receiver import *
+from leads_gpio.led import *
+from leads_gpio.led_group import *
