+++ leads_gpio/types.py
